requirements.txt
slack_sdk
flask
python-dotenv
cryptography
requests